// qwikJsLearning2023/simpsons-app/src/types/type.d.ts
// FILE: types/type.d.ts
// _______________________________________________

export type CharacterType = {
  quote: string
  character: string
  image: string
  characterDirection: string
}
// _______________________________________________

export type CharacterListResponse = Array<CharacterType>
// _______________________________________________